+++ src/main/java/com/mongodb/kafka/connect/sink/writemodel/strategy/DefaultWriteModelStrategy.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2008-present MongoDB, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Original Work: Apache License, Version 2.0, Copyright 2017 Hans-Peter Grahsl.
+ */
+
+package com.mongodb.kafka.connect.sink.writemodel.strategy;
+
+import org.bson.BsonDocument;
+
+import com.mongodb.client.model.WriteModel;
+
+import com.mongodb.kafka.connect.sink.Configurable;
+import com.mongodb.kafka.connect.sink.MongoSinkTopicConfig;
+import com.mongodb.kafka.connect.sink.converter.SinkDocument;
+
+public class DefaultWriteModelStrategy implements WriteModelStrategy, Configurable {
+
+  private WriteModelStrategy writeModelStrategy = new ReplaceOneDefaultStrategy();
+
+  @Override
+  public WriteModel<BsonDocument> createWriteModel(final SinkDocument document) {
+    assert writeModelStrategy != null;
+    return writeModelStrategy.createWriteModel(document);
+  }
+
+  @Override
+  public void configure(final MongoSinkTopicConfig configuration) {
+    if (configuration.isTimeseries()) {
+      writeModelStrategy = new InsertOneDefaultStrategy();
+    }
+  }
+
+  WriteModelStrategy getWriteModelStrategy() {
+    return writeModelStrategy;
+  }
+}